Rename spiutils::driver to spiutils::driver::spi_device.
